kepctl: Move command logic into pkg/
