+++ pkg/kepctl/commands/promote.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package cmd
+package commands
 
 import (
 	"github.com/pkg/errors"
